+++ tests/Unit/DataRepository/CephFileDataRepositoryTest.php
@@ -259,11 +259,7 @@ class CephFileDataRepositoryTest extends TestCase
         $ret = $this->sut->findAll();
 
         $expected = $this->data;
-        foreach ($expected as &$data) {
-            if (!isset($data['Metadata'])) {
-                $data['Metadata'] = [];
-            }
-        }
+        $this->replaceEmptyMetadataByArray($expected);
 
         $this->assertEquals($expected, $ret);
     }
@@ -375,12 +371,7 @@ class CephFileDataRepositoryTest extends TestCase
     public function testFindBy(array $criteria, ?int $limit, bool $continue, array $expectedResult, ?array $expectedContinueResult = null): void
     {
         $ret = $this->sut->findBy($criteria, null, $limit, (int)$continue);
-
-        foreach ($expectedResult as &$data) {
-            if (!isset($data['Metadata'])) {
-                $data['Metadata'] = [];
-            }
-        }
+        $this->replaceEmptyMetadataByArray($expectedResult);
 
         $this->assertEquals($ret, $expectedResult);
 
@@ -389,14 +380,42 @@ class CephFileDataRepositoryTest extends TestCase
         }
 
         $ret = $this->sut->findBy($criteria, null, $limit, 1);
-
-        foreach ($expectedContinueResult as &$data) {
-            if (!isset($data['Metadata'])) {
-                $data['Metadata'] = [];
-            }
-        }
+        $this->replaceEmptyMetadataByArray($expectedContinueResult);
 
         $this->assertEquals($ret, $expectedContinueResult);
+    }
+
+    public function providerFindByWithOrderBy(): array
+    {
+        return [
+            [['bucket' => -1], [2, 3, 4, 0, 1]],
+            [['id' => 1], [0, 2, 1, 3, 4]],
+            [['id' => 1, 'bucket' => -1], [2, 0, 3, 1, 4]],
+            [['bucket' => -1, 'id' => -1], [4, 3, 2, 1, 0]],
+            [['metadata' => ['mymetadata2' => -1]], [3, 4, 2, 0, 1]],
+            [['metadata' => ['mymetadata1' => 1]], [1, 3, 0, 4, 2]],
+            [['metadata' => ['mymetadata1' => 1], 'bucket' => -1], [3, 1, 4, 0, 2]],
+            [['bucket' => -1, 'metadata' => ['mymetadata2' => -1]], [3, 4, 2, 0, 1]]
+        ];
+    }
+
+    /**
+     * @dataProvider providerFindByWithOrderBy
+     *
+     * @covers ::findBy
+     * @covers ::sort
+     */
+    public function testFindByWithOrderBy(array $orderBy, array $expectedResultKeys): void
+    {
+        $expectedResult = [];
+        foreach ($expectedResultKeys as $key) {
+            $expectedResult[] = $this->data[$key];
+        }
+
+        $ret = $this->sut->findBy([], $orderBy);
+        $this->replaceEmptyMetadataByArray($expectedResult);
+
+        $this->assertEquals($ret, $expectedResult);
     }
 
     public function providerFindByWithIdAndContinueShouldThrowException(): array
@@ -511,6 +530,20 @@ class CephFileDataRepositoryTest extends TestCase
     }
 
     /**
+     * Add an empty array as metadata to expected results if metadata is missing
+     *
+     * @param File[] $result expected result
+     */
+    private function replaceEmptyMetadataByArray(array &$result): void
+    {
+        foreach ($result as &$data) {
+            if (!isset($data['Metadata'])) {
+                $data['Metadata'] = [];
+            }
+        }
+    }
+
+    /**
      * @dataProvider providerFindByFrom
      *
      * @covers ::findByFromCalled
@@ -521,11 +554,7 @@ class CephFileDataRepositoryTest extends TestCase
         $this->sut->findByFromCalled($criteria, $from, null, null);
         $ret = $this->sut->findBy($criteria);
 
-        foreach ($expectedResult as &$data) {
-            if (!isset($data['Metadata'])) {
-                $data['Metadata'] = [];
-            }
-        }
+        $this->replaceEmptyMetadataByArray($expectedResult);
 
         $this->assertEquals($ret, $expectedResult);
     }