+++ src/EventListener/LazyLoadedProperyGetListener.php
@@ -3,9 +3,6 @@
 
 namespace Coffreo\CephOdm\EventListener;
 
-
-use Coffreo\CephOdm\Entity\File;
-
 interface LazyLoadedProperyGetListener
 {
     public function lazyLoadedPropertyGet($object, string $propertyName): void;